Added a bunch of tests
